# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment Service API",
        "description": "API for Payment service project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/cmi/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "redirect request from CMI",
                "operationId": "App\\Http\\Controllers\\V1\\CmiController::redirto",
                "parameters": [
                    {
                        "name": "t_id",
                        "in": "query",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "10000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/fawry/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "background callback from fawry",
                "operationId": "App\\Http\\Controllers\\V1\\FawryController::redirto",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/globaliris/redirto": {
            "get": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from globaliris",
                "operationId": "App\\Http\\Controllers\\V1\\GlobalirisController::redirto",
                "parameters": [
                    {
                        "name": "t_id",
                        "in": "query",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "10000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/pay_later/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from paygate",
                "operationId": "App\\Http\\Controllers\\V1\\PayLaterController::redirto",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/paypal/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from paygate",
                "operationId": "App\\Http\\Controllers\\V1\\PayPalController::redirto",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/paygate/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from globaliris",
                "operationId": "App\\Http\\Controllers\\V1\\PaygateController::redirto",
                "parameters": [
                    {
                        "name": "t_id",
                        "in": "query",
                        "description": "transaction id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "10000"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
        "/api/v1/tingg/redirto": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from paygate",
                "operationId": "App\\Http\\Controllers\\V1\\TinggController::redirto",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
        "/api/v1/tingg/return": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from paygate",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
        "/api/v1/tingg/notify": {
            "post": {
                "tags": [
                    "Payment API"
                ],
                "description": "return reqeust from paygate",
                "operationId": "App\\Http\\Controllers\\V1\\TinggController::notify",
                "responses": {
                    "200": {
                        "description": "transaction created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Error: bad request"
                    }
                }
            }
        },
...
    }
}